[Feat] 주문 요청 비지니스 로직 구현, 테스트 코드 작성 완료

+++ src/main/java/com/barogo/api/domain/order/entity/Order.java
@@ -42,7 +42,7 @@ public class Order extends DateEntity {
     private String productName;
 
     @Column(name = "product_count")
-    private BigDecimal productCount;
+    private Integer productCount;
 
     @Column(name = "product_unit_price")
     private BigDecimal productUnitPrice;
@@ -50,7 +50,7 @@ public class Order extends DateEntity {
     @Builder
     public Order(LocalDateTime orderDate, String address, OrderStatus status,
                  LocalDateTime completeDate, BigDecimal amount, BigDecimal fee,
-                 String productName, BigDecimal productCount, BigDecimal productUnitPrice) {
+                 String productName, Integer productCount, BigDecimal productUnitPrice) {
         this.orderDate = orderDate;
         this.address = address;
         this.status = status;
